backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
fairlearn==0.10.0
aif360==0.6.1
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
google-cloud-bigquery==3.23.1
google-cloud-firestore==2.16.0
google-cloud-aiplatform==1.57.0
requests==2.32.2
pydantic==2.7.1
python-dotenv==1.0.1
pyyaml==6.0.1
